requirements.txt
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.9
Flask-RESTX==1.3.0
python-dotenv==1.0.0
marshmallow==3.21.1
Flask-Marshmallow==0.15.0
marshmallow-sqlalchemy
requests
hdfs